axo q já podemos dar fix #1 por encerrado, todos os testes estão passando corrigido retorno expr para double e demais locais relacionados

+++ mycalc.h
@@ -9,5 +9,5 @@ int exit_with_error(int);
 
 extern token_t gettoken(FILE *);
 
-extern int expr(void);
+extern double expr(void);
